Fix lists (#3)
When a user type or a raw Dict are uploaded their fields are parsed into a root typed JSON Map. When they are downloaded from DynamoDB, the fields are provided directly (and placed back into a Map when parsed). Previously, only non-typed Dicts were re-Mapped, leading to an error when returning certain object types. Tests updated to respect the Mapped nature of Typed objects.

* Objects and Dicts are maps
* Update tests